requirements.txt
Pillow
pillow-avif-plugin
